requirements.txt
streamlit
google-generativeai
python-docx
PyPDF2
pandas
python-dotenv
scikit-learn
sentence-transformers
